mysql-mcp/requirements.txt
fastmcp>=0.1.0
mysql-connector-python>=8.0.0
